fix: staking contract countdown (#520)
* Fix: return NA instead of message when time has passed

* refactor: remove unused time formatting functions

* feat: implement countdown for staking contract migration and refactor related components
