+++ frontend/utils/time.ts
@@ -1,11 +1,5 @@
 import { isNumber } from 'lodash';
 
-import { NA } from '@/constants/symbols';
-
-const MILLISECONDS_IN_A_DAY = 24 * 3600 * 1000;
-const MILLISECONDS_IN_AN_HOUR = 3600 * 1000;
-const MILLISECONDS_IN_A_MINUTE = 60 * 1000;
-
 export const getTimeAgo = (timestampInSeconds: number) => {
   if (!isNumber(timestampInSeconds)) return null;
 
@@ -70,33 +64,4 @@ export const formatToShortDateTime = (timeInMs?: number) => {
     hour12: true,
     timeZone: 'UTC',
   });
-};
-
-/**
- * @returns formatted time remaining in the format of 'X days Y hours Z minutes'
- * @example 1626825600 => '1 day 2 hours 30 minutes'
- */
-export const formatTimeRemainingFromNow = (
-  futureTimestampInSeconds: number,
-) => {
-  if (!isNumber(futureTimestampInSeconds)) return NA;
-
-  const now = new Date().getTime();
-  const targetTime = futureTimestampInSeconds * 1000 + Date.now();
-  const timeDifference = targetTime - now;
-
-  if (timeDifference <= 0) return 'Time has passed';
-
-  const days = Math.floor(timeDifference / MILLISECONDS_IN_A_DAY);
-  const hours = Math.floor(
-    (timeDifference % MILLISECONDS_IN_A_DAY) / MILLISECONDS_IN_AN_HOUR,
-  );
-  const minutes = Math.floor(
-    (timeDifference % MILLISECONDS_IN_AN_HOUR) / MILLISECONDS_IN_A_MINUTE,
-  );
-
-  const daysInWords = `${days} day${days !== 1 ? 's' : ''}`;
-  const hoursInWords = `${hours} hour${hours !== 1 ? 's' : ''}`;
-  const minutesInWords = `${minutes} minute${minutes !== 1 ? 's' : ''}`;
-  return `${daysInWords} ${hoursInWords} ${minutesInWords}`.trim();
 };
